Add support for:   --ftp-passive,   --ftp-regular,   --num-retries
Removed -m option on FTP put command.  This means that
the remote directory must exist prior to backup.

Changed ftpBackend from -f option back to commandline.
Various versions of ncftp* interact differently when
both -f and commandline options are supplied.

The FTP password is munged in all log operations.

Added logging of filenames in the bucket when -v9 is
used on Amazon S3.