Add support for Shared Access Signatures to the Azure backend.
Authenticating to the Azure storage via the account is suboptimal as it
grants the process full administrative permissions on the storage
account. Usage of a shared access signature allows to pass only the
minimal permissions on a single container to Duplicity. This makes it
much more sutable for automated usage, e.g. in cron jobs.