Wait for Glacier batch unfreeze to finish
The ThreadPoolExecutor starts the unfreezing of volumes in parallel.
However we can wait until it finishes its work for all volumes.

This currently does _not_ wait until the unfreezing process has
finished, but only until the S3 'restore()' operations have finished
(which can take a bit time).

The actual (sequential) pre_processing of the volumes to restore then
waits for the actual unfreezing to finish by regularly checking the
state of the unfreezing.