Avoid calling stat when checking for exclude-if-present files
If a folder with rw- permissions (i.e. read and write, but no exec)
is examined for the presence of an exclude-if-present file, we would
previously throw an exception when trying to stat the file during
Path object construction.

But we don't need to stat in this case. This patch just calls
listdir() and checks if the file is in that result.